nonogram logic divided into two classes for rows and columns
