+++ puzzleAppJavaBackend2.0/puzzleAppFileManager/src/main/java/com/puzzlesolverappbackend/puzzleAppFileManager/NonogramSolver.java
@@ -44,7 +44,8 @@ public class NonogramSolver {
 
         this.rootNode = gsonObj.fromJson(gsonObj.toJson(this.solutionNode), NonogramSolutionNode.class);
 
-        this.finalSolutionLogic = new NonogramLogic(nonogramLogic.getRowsSequences(), nonogramLogic.getColumnsSequences());
+        this.finalSolutionLogic = new NonogramLogic(nonogramLogic.getRowsSequences(),
+                nonogramLogic.getColumnsSequences());
 
         // optional correct solution
         this.solutionFileName = "r" + fileName;
