+++ puzzleAppJavaBackend2.0/puzzleAppFileManager/src/main/java/com/puzzlesolverappbackend/puzzleAppFileManager/payload/NonogramActions.java
@@ -1,0 +1,4 @@
+package com.puzzlesolverappbackend.puzzleAppFileManager.payload;
+
+public class NonogramActions {
+}
