+++ puzzleAppJavaBackend2.0/puzzleAppFileManager/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.sql.init.platform=mysql
 spring.datasource.url= jdbc:mysql://localhost:3306/puzzles?useSSL=false&serverTimezone=UTC
 spring.datasource.username= root
-spring.datasource.password=ABCDE
+spring.datasource.password=HubZdziuBD7aM?
 
 spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.MySQL5InnoDBDialect
